decoiupling sparse and dense ingest pipeline integration tests
